services/feature-extraction/requirements.txt
kafka-python==2.0.2
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
python-dotenv==1.0.0
